Add ShouldProcess support to functions

+++ RDCMan/Public/Close-RDCMan.ps1
@@ -1,4 +1,5 @@
 function Close-RDCMan {
+    [CmdletBinding(SupportsShouldProcess)]
     param (
         [switch]
         $Force
@@ -7,13 +8,18 @@ function Close-RDCMan {
     $Running = Get-Process -Name RDCMan -ErrorAction SilentlyContinue
 
     foreach ($Process in $Running) {
-        $null = $Process.CloseMainWindow()
+        # Attempt to close gracefully
+        if ($PSCmdlet.ShouldProcess("RDCMan window [$($Process.Id)] with title $($Process.MainWindowTitle)", 'Close Window')) {
+            $null = $Process.CloseMainWindow()
+        }
 
         # Forcefully close remaining processes, if specified
         if ($Force -and (Get-Process -Id $Process.Id -ErrorAction SilentlyContinue)) {
             Write-Verbose "Process [$($Process.Id)] with title $($Process.MainWindowTitle) did not close gracefully, stopping [$($Process.Id)"
 
-            Stop-Process -Id $Process.Id -Force
+            if ($PSCmdlet.ShouldProcess("RDCMan process [$($Process.Id)] with title $($Process.MainWindowTitle)", 'Stop Process')) {
+                Stop-Process -Id $Process.Id -Force
+            }
         }
     }
 }
